Move expose engine to app exposer
