+++ forguarder/packages/api/src/appExposer/HttpsMappingParser.ts
