+++ forguarder/packages/api/src/appExposer/HttpsMappingReader.ts
