+++ forguarder/packages/api/src/server/routes/routers/apps/handlers/updateApps.ts
@@ -1,5 +1,5 @@
 import { Logger } from "../../../../../logger/index.js";
-import { AppsSettings } from "../../../../../exposureSettingsEngine/types.js";
+import { AppsSettings } from "../../../../../appExposer/types.js";
 
 export const updateAppsHandler = async ({ input }: { input: AppsSettings }): Promise<void> => {
 	// TODO: Do things with input
