+++ forguarder/packages/api/src/server/routes/routers/apps/index.ts
@@ -1,7 +1,7 @@
 import { publicProcedure, router } from "../../trpc.js";
 import { getAppsHandler } from "./handlers/getApps.js";
 import { updateAppsHandler } from "./handlers/updateApps.js";
-import { AppsSettingsSchema } from "../../../../exposureSettingsEngine/schemas.js";
+import { AppsSettingsSchema } from "../../../../appExposer/schemas.js";
 
 export const appsRoutes = router({
 	getApps: publicProcedure.output(AppsSettingsSchema).query(getAppsHandler),
